APIs related to completeAccountStatus moved to AccountController
