+++ backend/src/Service/StoreOwner/StoreOwnerProfileService.php
@@ -6,7 +6,7 @@ use App\AutoMapping;
 use App\Constant\StoreOwner\StoreProfileConstant;
 use App\Constant\User\UserReturnResultConstant;
 use App\Entity\StoreOwnerProfileEntity;
-use App\Request\Main\CompleteAccountStatusUpdateRequest;
+use App\Request\Account\CompleteAccountStatusUpdateRequest;
 use App\Request\StoreOwner\StoreOwnerProfileStatusUpdateByAdminRequest;
 use App\Request\StoreOwner\StoreOwnerProfileUpdateByAdminRequest;
 use App\Request\StoreOwner\StoreOwnerProfileUpdateRequest;
